+++ pyked/chemked.py
@@ -152,6 +152,13 @@ class ChemKED(object):
                 Only the first author is printed when ``Reference`` or ``Reference:Authors`` is
                 selected because the whole author list may be quite long.
 
+        Note:
+            If the Composition is selected as an output type, the composition specified in the
+            `DataPoint` is used. No attempt is made to convert to a consistent basis; mole fractions
+            will remain mole fractions, mass fractions will remain mass fractions, and mole percent
+            will remain mole percent. Therefore, it is possible to end up with more than one type of
+            composition specification in a given column. Caveat Emptor.
+
         Examples:
             >>> df = ChemKED(yaml_file).get_dataframe()
             >>> df = ChemKED(yaml_file).get_dataframe(['Temperature', 'Ignition Delay'])
@@ -200,11 +207,14 @@ class ChemKED(object):
         data = []
         for d in self.datapoints:
             row = []
+            d_species = [s['species-name'] for s in d.composition]
             for col in col_labels:
                 if col in species_list:
-                    for s in d.composition:
-                        if col == s['species-name']:
-                            row.append(s['amount'])
+                    if col in d_species:
+                        s_idx = d_species.index(col)
+                        row.append(d.composition[s_idx]['amount'])
+                    else:
+                        row.append(Q_(0.0, 'dimensionless'))
                 elif 'reference' in col or 'apparatus' in col:
                     split_col = col.split(':')
                     if split_col[1] == 'authors':